split blackberry.io.dir into webworks.handset and tablet platforms, created SDCard and store directories on startup
