+++ lib/ripple/platform/webworks.core/2.0.0/client/io/dir.js
@@ -41,20 +41,10 @@ module.exports = {
             post: {path: _prunePrefix(path)}
         });
     },
-    getFreeSpaceForRoot: function (path) {
-        return transport.call(_uri + "getFreeSpaceForRoot", {
-            post: {path: _prunePrefix(path)}
-        });
-    },
     getParentDirectory: function (path) {
         return _addPrefix(transport.call(_uri + "getParentDirectory", {
             post: {path: _prunePrefix(path)}
         }));
-    },
-    getRootDirs: function () {
-        return transport.call(_uri + "getRootDirs", {}).map(function (dir) {
-            return FILE + dir;
-        });
     },
     listDirectories: function (path) {
         return transport.call(_uri + "listDirectories", {
