+++ lib/ripple/platform/webworks.core/2.0.0/fsCache.js
@@ -24,7 +24,12 @@ var fs = require('ripple/fs'),
     _self;
 
 event.on("FileSystemInitialized", function () {
-    _self.refresh();
+    jWorkflow.order(function (prev, baton) {
+        baton.take();
+        fs.mkdir("/SDCard", function () {
+            fs.mkdir("/store", baton.pass, baton.pass);
+        }, baton.pass);
+    }).start(_self.refresh);
 });
 
 function _fsError(e) {
