+++ lib/ripple/platform/webworks.core/2.0.0/server/io/dir.js
@@ -34,14 +34,8 @@ module.exports = {
     exists: function (get, post, baton) {
         return _packet(cache.dir.exists(post.path));
     },
-    getFreeSpaceForRoot: function (get, post, baton) {
-        return _packet(cache.dir.getFreeSpaceForRoot(post.path));
-    },
     getParentDirectory: function (get, post, baton) {
         return _packet(cache.dir.getParentDirectory(post.path));
-    },
-    getRootDirs: function (get, post, baton) {
-        return _packet(cache.dir.getRootDirs());
     },
     listDirectories: function (get, post, baton) {
         return _packet(cache.dir.listDirectories(post.path));
