+++ lib/ripple/platform/webworks.handset/2.0.0/server.js
@@ -50,7 +50,7 @@ module.exports = {
         },
         phone: require(platform + "phone"),
         io: {
-            dir: require(core + "io/dir"),
+            dir: require(platform + "io/dir"),
             file: require(core + "io/file")
         }
     }
