+++ lib/ripple/platform/webworks.handset/2.0.0/spec.js
@@ -271,7 +271,7 @@ module.exports = {
                 io: {
                     children: {
                         dir: {
-                            path: "webworks.core/2.0.0/client/io/dir",
+                            path: "webworks.handset/2.0.0/client/io/dir",
                             feature: "blackberry.io.dir"
                         },
                         file: {
